feat(events): extract noetl-events workspace crate (EE-4 PR 1)
First of three rounds on EE-4 (noetl/ai-meta#49 follow-up).  Carves
the shared event-envelope types out of `noetl-executor::events`
into a new dependency-light `noetl-events` workspace crate so the
noetl-server can link the canonical wire shape directly rather
than maintaining a hand-aligned `EventRequest` type that drifts
every time a field is added.

What this round does:

* Adds a new `events/` workspace member with `noetl-events 0.1.0`.
  Moves `ExecutorEvent` + `EventSink` + `EventEmitter` + `NoopSink`
  + every related test verbatim from `executor/src/events.rs` into
  `events/src/lib.rs`.  Crate-level docs describe the wire-format
  alignment with the Python `EventEmitRequest` shape.
* `executor/src/events.rs` becomes a thin re-export
  (`pub use noetl_events::{...}`) so every existing call site in
  the cli's local-mode runner AND in noetl-worker
  (`pub use noetl_executor::events::ExecutorEvent`) keeps
  compiling and running unchanged.
* `executor/Cargo.toml` gains `noetl-events = { version = "0.1.0",
  path = "../events" }`.  Once `noetl-events` ships on crates.io
  (EE-4 PR 2) this can flip to a versioned dep with a fallback
  path.
* No version bump on `noetl-executor` yet — its public API is
  source-compatible via the re-export, and the published 0.3.1
  on crates.io continues to carry the inlined types.  EE-4 PR 2
  bumps to 0.4.0 alongside the crates.io publish of noetl-events.

What this round does NOT do:

* Crates.io publish (EE-4 PR 2).
* noetl-server adoption — replacing the hand-aligned `EventRequest`
  with a thin wrapper around `noetl_events::ExecutorEvent` (EE-4
  PR 3).

Tests: 6/6 noetl-events tests pass (the four envelope-shape
contracts: optional fields omit-when-none / serialize-when-some,
`payload` → `context` alias, missing-optional roundtrip).  Full
noetl-executor lib suite stays green (102/102 — no source changes
to non-events modules).  Cli binary builds.

Refs noetl/ai-meta#49
